backend/pyproject.toml
[project]
name = "memagent-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agno>=2.4.8",
    "aiosqlite>=0.22.1",
    "alembic>=1.18.3",
    "asyncpg>=0.31.0",
    "fastapi>=0.128.4",
    "google-api-python-client>=2.189.0",
    "google-auth>=2.48.0",
    "google-auth-oauthlib>=1.2.4",
    "piexif>=1.1.3",
    "pillow>=12.1.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dateutil>=2.9.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.46",
    "structlog>=25.5.0",
    "tiktoken>=0.12.0",
    "unicorn[standard]>=2.1.4",
    "uvicorn[standard]>=0.40.0",
    "websockets>=13.0,<15.1.0",
    "google-genai>=1.62.0",
    "google-generativeai>=0.8.6",
    "httpx>=0.28.1",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
